physics-bh: Use C11 atomics for shared vars, free octrees upon exit
Yes, it's decimate, not desecrate, no typo here. I have an idea how to thread the octree generation, might be clumsy but I'll give it a go. Also, using C11 atomics instead of a mutex because we already require it(as --std=gnu11). Atomics definately seem the most interesting feature of C11. Pthreads already does everything you need threads to do(except sending signals, that's still clumsy). Also, lua_tonumber always returns double. And the "by specifications" part means I have let to lay my eyes upon a coordinate taking more than 10 decimal digits in any kind of a format, so might as well stick to floats. Also, MiB > MB. Literally and figuratively speaking. Over and out!

+++ src/options.h
@@ -22,7 +22,7 @@ struct option_struct* option;
 
 struct option_struct {
 	float dt, bh_ratio;
-	long double elcharge, gconst, epsno;
+	double elcharge, gconst, epsno;
 	unsigned int obj;
 	unsigned short int avail_cores, verbosity;
 	int width, height;
